Add PopulationSpace component that displays 2 side-by-side panels
Add a PopulationSpace component that displays 2 side-by-side panels.  Remove the now obsolete InvestigationPanel component.  Display the PopulationSpace component when the user clicks on the population button in the investigation dialog.  Update the UI model to contain population graph state (this will need to be reworked to have greater complexity.  We could use an enum that reflects the current space that is visible.  This is probably best done in its own story and branch).
[#161766223]