Create authoring for the populations model
This adds to the authoring schema a set of properties nested under the keyword
"populations", which is passed as-is to the populations store.

(The store then passes it off to its internal curriculum-specific store, but
this is an implementation detail.)

In order to keep flat MST model properties, yet allow for the richer format of
nested authoring properties, the properties under "population" are flattened
before being used to create the store.

This also moves the default properties into the `defaultAuthoring` object
instead of having them in the url-parameters objects.

[#161658945]