Add additional spaces to UI model and allow them to be shown in UI
Add additional spaces to UI model and allow them to be shown in UI.  At present, I added a "none" space to indicate no space is shown.  Not entirely clear if this is desired or even possible, but allows no space to be shown.  Also note that for the breeding and organism spaces, we're just showing the population space until these other spaces are built
[#161784877]