Make two receptors work correctly for all mice
For brown mice, both receptors work, for tan only one does, and for
white neither does.

This requires some ugly duplication of state for the g-proteins and
hormones. Ever state needs to decide whether it's acting in relation
to the first or second receptor. A future version of Organelle should
improve this such that the agents can communicate better with their
nearest other agents, and set state on them instead of the world state.

[#162053547]
[#161427499]
[#161826609]