Use numerical value in model to determine color
Instead of the two properties `working_receptor` and `tan`, we now have
a single `base_darkness` property, which works well because
`if base_darkness` is essentially the same as the old `if working_receptor`.

Also it is a limitation of Organelle that we can't currently check rules
against string values (e.g. `if color === "tan"`), so using numerical
values is good for this reason as well.