Load organisms space model
This serializes the Organisms model on save, and creates the Organisms
model on load.

Because the Organisms model contains `type.reference` references to
`BackpackMouse`, and because the trees are separate, when we create the
Organisms model naively we can't find the references BackpackMice.
Therefore we pre-process the serialized data and find the referenced
mice before creating the Organisms model.

This creates the entire Organisms model, minus the transient `time` and 
`substanceDeltas` which were removed in the last commit, and minus other
transient props which are removed in postProcessSnapshot.

[#162664356]