Downsample data points when they grow too large.
This uses an algorithm	which attempts to preserve major features of the
data while downsampling.

We downsample the entire dataset every 40 datapoints, and then stream
the next 40 points as-is. This avoids downsampling every tick, which
causes the data to change every tick, which is visually jarring.

This also changes the line chart to only use `visibleData`.

[#163801519]