Scale x axis points, and add initialMaxA1 to charts
For the mice we want each step to be labeled as 10x the step number.

However, when we do this, the chart assumes we want our initial maximum
to still be the maxPoints (20), and won't update that maximum until we
pass 20 points, at which point we are already at y=200.

This adds a new prop, `initialMaxA1`, which hard-codes our desired initial
max, since there is no way for the dataset to be able to calculate it
otherwise.

[#161425423]