Add model props for disabling graph buttons
If `enable__Chart` is false, the button for that chart will be given the
disabled class and will not have an associated action.

This also requires making `chartType` into a private property with a
view, because it needs to be calculated late in case the color chart is
disabled. This further requires extracting the function to set up the
chart based on the color type, since we may be switching chart types
on the first load of the model.

[#163743063]