Add mouse to org space when user clicks on backpack
This changes the UX to avoid needing to click on the 4-up panel's collect
button after clicking on a backpack mouse. Instead, if the user clicks on
a backpack mouse while the org space is visible, and there is an empty row,
we will add the mouse to the first empty row.

This management between the stores is done in stores.ts, which is cleaner
than adding listeners in the views.

[#164258750]