Only show zoom-in gif if we're coming from org level
Previously, we wanted to show the zoom-in gif even if we were at a
deeper level if the cell hadn't finished loading yet. However, in the
case where we're starting at the nucleus level, the cell model will
never load, and our logic had us showing the organism zoom.

This changes things so that we only show that organism zoom if we're on
the organism level or we've just come from the organism level.