Save and load back in environment when switching spaces
We used to re-create the populations model each time we moved to a
different space. Now we just destroy and re-create the environment,
which is necessary because there are a number of listeners related to
the view that need to be destroyed and created afresh.

By recreating the environment within the same model, most properties
will simply be brought back in in the same state they were. We just
need to serialize the agents and the date, which are not mirrored in
the Mobx model.

[#171865782]