Return previous slider behavior if no pie charts
When pie-charts are visible, the slider behavior allows us to move the
pointer all the way back to zero, which means there is a section of the
slider track that does not move the graph.

When pie charts are hidden, this returns the previous behavior of having
the entire track represent the fraction of the chart we need to slide.