Don't create populations space for peas
If a unit has no model for populations, it doesn't make sense to create
the store.

This requires updating components to return null if the populations
store does not exist, though in some nested components we can safely
use `store.populations!`.

Populations Space is already nicely packaged into a unit-specific folder,
so nothing more needs to be done.