Don't create organisms store for peas
Making `organisms` optional is a much bigger change than `population`
as so many more components need to observe the organisms model.

Where sensible, components that expect `organisms` to exist return null
if it isn't there, degrading without throwing errors. To avoid using
`organisms!` everywhere, a number of methods now take organisms as an
argument.

On a few innermost components, we can use `organisms!` without fear, as
we can only be in those components if higher-level components are being
rendered, and the pattern of passing in organisms each time starts to
make the code very ugly.