Update parent and offspring images
With this optional image and subsequent ones, the logic of which image
to show resides in BackpackMouse, so the views can just access this
directly as a prop.

This is fairly messy, because which images are used in the zoomed-out
"nests" view, the parents, the offspring, and the charts, are basically
completely non-overlapping between mice and peas.