Custom inspect info panels for different species
This is the first time breaking out new JSX views for the different
species, but the amount of special-casing just made the info panels
too complex, and the species definitions too awkward.

On the one hand, it seems like it would be good to have each species
definition simply be declaratory, and avoid the need to write JSX for
every new species we introduce.

On the other hand, the species files may have been much more
straight-forward had I done this from the start...