Auto-switch to inspect mode when we view experiment
Also remove the behavior where clicking on the same button will switch
the user to "none" mode. This isn't a useful mode, and may confuse
users who are used to clicking Inspect after entering breeding view.

This means that, in the peas unit, a user will always be in inspect
mode while they are in the breeding view. In the mouse unit, a user
can switch between inspect and select modes.