Add start of very simple Formik authoring form
This is broken up so we can see a minimum-viable authoring form. This
updates just one property, but does so correctly (the submit button
opens a model with the correct settings).

Also, given an authored model, we can change index.html to mice-authoring.html
and see the authoring form with the authored properties.